put fscache in its own file for log-name
if other cache systems are added, it should be recognizable
in the logs which cache system logged a certain msg
